requirements.txt
streamlit==1.54.0
google-genai==1.64.0
qdrant-client
